move some computed methods of match overview in mixins
