+++ src/views/Creator.vue
@@ -157,7 +157,7 @@ export default {
 
   data() {
     return {
-      step: 2,
+      step: 1,
       snackbar: false,
     };
   },
